+++ docs/changes.md
@@ -1,5 +1,8 @@
 # 最新动态
 
+20223/01/05
+  * 完善 tab button和pages(感谢俊杰提供补丁)
+
 20223/01/04
   * widget为designing\_window的子控件时，widget\_set\_need\_relayout失败的问题(感谢朝泽提供补丁)
   * 修复了fs\_create\_dir\_r函数中tokenizer没有释放就return的问题(感谢泽武提供补丁)